split E2E tests out and fan out E2Es to run in parallel and allow for automatic discoverability
