+++ .github/workflows/_buildDeploy.yml
@@ -50,15 +50,8 @@ jobs:
       - name: Deploy FLEX AWS infra
         run: cd ./platform/infra/flex && pnpm run deploy
 
-      - name: E2E tests
+      - name: Platform E2E tests
         if: inputs.STAGE != 'production'
         env:
           STAGE: ${{ env.STAGE }}
-        run: pnpm --filter @flex/e2e test:e2e
-
-      - name: E2E tests (domain)
-        if: inputs.STAGE != 'production'
-        continue-on-error: true
-        env:
-          STAGE: ${{ env.STAGE }}
-        run: pnpm --filter @flex/e2e test:e2e:all
+        run: pnpm --filter @flex/e2e test:e2e:platform
