+++ .github/workflows/_moduleE2E.yml
@@ -1,0 +1,84 @@
+name: Module E2E
+
+on:
+  workflow_call:
+    inputs:
+      AWS_REGION:
+        description: "AWS Region to run against"
+        type: string
+      STAGE:
+        description: "Development stage"
+        type: string
+        required: true
+    secrets:
+      ROLE_TO_ASSUME:
+        description: "AWS role to assume"
+        required: true
+
+jobs:
+  discover:
+    name: Discover module E2E suites
+    runs-on: ubuntu-latest
+    outputs:
+      modules: ${{ steps.list.outputs.modules }}
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
+
+      - name: List domains with a test:e2e script
+        id: list
+        # Emit a JSON array of { name, package } for every domain that declares
+        # a test:e2e script. This is the discovery contract: adding a domain with
+        # a test:e2e script produces its own check with no edit to this workflow.
+        run: |
+          modules=$(node -e '
+            const fs = require("fs"), path = require("path");
+            const out = [];
+            for (const d of fs.readdirSync("domains")) {
+              const pj = path.join("domains", d, "package.json");
+              if (!fs.existsSync(pj)) continue;
+              const j = JSON.parse(fs.readFileSync(pj, "utf8"));
+              if (j.scripts && j.scripts["test:e2e"]) out.push({ name: d, package: j.name });
+            }
+            process.stdout.write(JSON.stringify(out));
+          ')
+          echo "modules=$modules" >> "$GITHUB_OUTPUT"
+
+  test:
+    name: Module E2E (${{ matrix.module.name }})
+    needs: discover
+    if: needs.discover.outputs.modules != '[]'
+    runs-on: ubuntu-latest
+    permissions:
+      id-token: write # Required for OIDC authentication with AWS
+    strategy:
+      fail-fast: false # one domain failing must not cancel the others
+      matrix:
+        module: ${{ fromJSON(needs.discover.outputs.modules) }}
+    env:
+      STAGE: ${{ inputs.STAGE }}
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
+
+      - name: Setup node
+        uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
+        with:
+          node-version-file: ".nvmrc"
+
+      - name: Install dependencies
+        run: |
+          npm install -g pnpm@latest-10
+          pnpm install
+
+      - name: Configure AWS credentials
+        uses: aws-actions/configure-aws-credentials@e7f100cf4c008499ea8adda475de1042d6975c7b # v6.2.0
+        with:
+          role-to-assume: ${{ secrets.ROLE_TO_ASSUME }}
+          aws-region: ${{ inputs.AWS_REGION }}
+          role-session-name: ${{ github.run_id }}
+
+      - name: Run module E2E tests
+        env:
+          STAGE: ${{ env.STAGE }}
+        run: pnpm --filter ${{ matrix.module.package }} test:e2e
