+++ .github/workflows/main.yml
@@ -27,10 +27,33 @@ jobs:
     secrets:
       ROLE_TO_ASSUME: ${{ secrets.DEV_DEPLOYMENT_ROLE }}
 
+  moduleE2EDev:
+    name: Module E2E (Development)
+    # Leaf job: runs after the development deploy but nothing depends on it, so a
+    # module failure leaves the run red while the deploy chain proceeds.
+    needs: deployDev
+    uses: ./.github/workflows/_moduleE2E.yml
+    with:
+      AWS_REGION: "eu-west-2"
+      STAGE: "development"
+    secrets:
+      ROLE_TO_ASSUME: ${{ secrets.DEV_DEPLOYMENT_ROLE }}
+
   deployStaging:
     name: Deploy to Staging
     needs: deployDev
     uses: ./.github/workflows/_buildDeploy.yml
+    with:
+      AWS_REGION: "eu-west-2"
+      STAGE: "staging"
+    secrets:
+      ROLE_TO_ASSUME: ${{ secrets.STAGING_DEPLOYMENT_ROLE }}
+
+  moduleE2EStaging:
+    name: Module E2E (Staging)
+    # Leaf job: runs after the staging deploy but nothing depends on it.
+    needs: deployStaging
+    uses: ./.github/workflows/_moduleE2E.yml
     with:
       AWS_REGION: "eu-west-2"
       STAGE: "staging"
