+++ domains/udp/tsconfig.json
@@ -1,6 +1,12 @@
 {
   "extends": "@flex/config/tsconfig.json",
-  "include": ["src/**/*.ts", "vitest.config.ts", "domain.config.ts"],
+  "include": [
+    "src/**/*.ts",
+    "e2e/**/*.ts",
+    "vitest.config.ts",
+    "vitest.e2e.config.ts",
+    "domain.config.ts"
+  ],
   "compilerOptions": {
     "paths": {
       "@domain": ["./domain.config.ts"],
