+++ domains/uns/tsconfig.json
@@ -8,5 +8,11 @@
       "@tests/fixtures": ["./src/tests/fixtures.ts"]
     }
   },
-  "include": ["src/**/*.ts", "vitest.config.ts", "domain.config.ts"]
+  "include": [
+    "src/**/*.ts",
+    "e2e/**/*.ts",
+    "vitest.config.ts",
+    "vitest.e2e.config.ts",
+    "domain.config.ts"
+  ]
 }
