+++ domains/uns/vitest.config.ts
@@ -1,10 +1,11 @@
 import { config } from "@flex/config/vitest";
-import { defineConfig, mergeConfig } from "vitest/config";
+import { configDefaults, defineConfig, mergeConfig } from "vitest/config";
 
 export default mergeConfig(
   config,
   defineConfig({
     test: {
+      exclude: [...configDefaults.exclude, "e2e/**"],
       setupFiles: ["@flex/testing/setup/sdk"],
       env: {
         AWS_REGION: "eu-west-2",
