+++ libs/testing/src/index.e2e.ts
@@ -1,5 +1,6 @@
 export type { E2EEnv } from "./config/env";
 export { e2eEnvSchema } from "./config/env";
+export { it } from "./e2e/it";
 export { extendIt } from "./extend/it.e2e";
 export type { ApiResponse } from "./fixtures/api";
 export { createApi } from "./fixtures/api";
