+++ tests/e2e/src/platform/private-gateway.test.ts
@@ -1,6 +1,5 @@
+import { it } from "@flex/testing/e2e";
 import { describe, expect } from "vitest";
-
-import { it } from "../extend/it";
 
 describe("private gateway", () => {
   it("rejects direct access from public internet (403 Forbidden)", async ({
